feat: Endless Engine v1.1.0 — complete idle game toolkit
Core systems (48 total): economy, generators, passive income, prestige,
ascension, research, upgrade/skill trees, merge, wave/combat, click loop,
harvest loop, buildings, pets, events, quests, achievements, leaderboard,
save/load with migrations, offline earnings, telemetry, Steam integration.

Sample scenes (6): MinimalIdle, ClickerIdle, HarvestLoop, WaveIdle,
MergeIdle, PrestHeavy — all scene warnings resolved, GUIDs verified.

Save system: atomic write, HMAC-SHA256 signing, dual-file backup,
migration pipeline V1→V4, crash-safe prestige snapshots.

Steam: Steamworks.NET integrated, achievements, cloud saves, leaderboards.

Tests: 83 test files, unit + integration + performance coverage.

Distribution: .gitignore added, Library/ untracked, com.unity.ugui
dependency declared in package.json, MinimalIdle sample imported.

+++ Assets/Tests/integration/full-system/integration-003-building-economy-chain_test.cs
@@ -99,7 +99,7 @@ namespace EndlessEngine.Tests.Integration.FullSystem
         {
             _economy.AddResources(500);
             _buildingService.TryPlace("test_farm", 0, 0);
-            long goldAfterPlace = _economy.CurrentResources; // 400
+            double goldAfterPlace = _economy.CurrentResources; // 400
 
             _buildingService.OnTick(1f); // 1 second tick
